Updated To Stripe 3.4.4 to allow for stripe connect
With Stripe 3.4.4 we now have the ability to use Stripe Connect which
is used for market places to charge a card and deposit funds into a
service providers account with one API call. These updates will allow
for these changes to take place. The required changes for
grovelabs/meteor-stripe-npm have already been committed and merged.